refactor: refactored all interfaces
