+++ src/category/category.interface.ts
@@ -1,0 +1,8 @@
+import { User } from '../user/user.schema';
+
+export interface createdCategoryDoc {
+  __v: string;
+  title: string;
+  color: string;
+  userId: User;
+}
