+++ src/image/image.interface.ts
@@ -1,0 +1,4 @@
+export interface UploadAvatar {
+  url: string;
+  public_id: string;
+}
