+++ src/notification/notification.interface.ts
@@ -1,0 +1,6 @@
+import { Types } from 'mongoose';
+
+export interface UserConnection {
+  userId: Types.ObjectId;
+  socketId: string;
+}
